Always define `start` and `stop` on `Service`
`Service`'s types say that `start` and `stop` can be undefined. This is
impossible in practice because the only way to instantiate a `Service`
is via `Registry.prototype.publish`, which always sets these properties.

This small refactor avoids this problem to ensure that `start` and
`stop` are always defined, which should make the library a bit easier to
use.